requirements.txt
aiohttp==3.6.2
aiohttp_cors==0.7.0
aiohttp_jwt==0.5.0
aiosqlite==0.11.0
asyncpg==0.20.1
asynctest==0.13.0
pytest-asyncio==0.10.0
tortoise-orm==0.15.11
uvloop==0.14.0
